+++ .scannerwork/report-task.txt
@@ -2,5 +2,5 @@ projectKey=restaurant_managent_system_react
 serverUrl=http://localhost:9000
 serverVersion=10.3.0.82913
 dashboardUrl=http://localhost:9000/dashboard?id=restaurant_managent_system_react
-ceTaskId=AYzZh_5gmrJHhJed6zwh
-ceTaskUrl=http://localhost:9000/api/ce/task?id=AYzZh_5gmrJHhJed6zwh
+ceTaskId=AY0WMCEmUS0tGi99CfnY
+ceTaskUrl=http://localhost:9000/api/ce/task?id=AY0WMCEmUS0tGi99CfnY
