--- src/components/Layout.css
@@ -1,3 +1,0 @@
-.Layout-wrapper{
-    background-color: rgba(192, 161, 89, 255);
-}
