+++ src/components/auth/form.css
@@ -16,7 +16,7 @@
     background-color: rgba(0, 67, 50, 255);
     border-style: solid;
     border-width: 1px;
-    border-color: rgba(192, 161, 89, 255);
+    border-color: white;
   }
 
 .form-elements {
@@ -37,11 +37,16 @@
     padding: 10px;
 }
 
-.button-design{
+.button-design-login-page{
   background-color: rgba(0, 67, 50, 255);
   border-width: 1px;
-  border-color: rgba(192, 161, 89, 255);
-  color: rgba(192, 161, 89, 255);
+  border-color: white;
+  color: white;
+}
+
+.button-design{
+  color: rgba(54,54,54,255);
+  border: 3px solid rgba(222,222,222,255);
 }
 
 .login-button{
@@ -51,7 +56,7 @@
 .form-title{
     align-self: center;
     margin-bottom: 50px;
-    color: rgba(192, 161, 89, 255);
+    color: white;
 }
 
 .error-message {
@@ -76,10 +81,10 @@
 
 .signup-text{
   align-self: center;
-  color: rgba(192, 161, 89, 255);
+  color: white;
 }
 
 .signup-link{
   text-decoration: none;
-  color: rgba(192, 161, 89, 255);
+  color: white;
 }
