+++ src/components/layout/Footer.css
@@ -1,7 +1,7 @@
 footer {
     background-color: rgba(0, 67, 50, 255);
     font-family: fangsong, serif;
-    color: rgba(192, 161, 89, 255);
+    color: white;
     display: flex;
     justify-content: center;
     align-items: center;
