+++ src/components/layout/Layout.css
@@ -1,0 +1,3 @@
+.Layout-wrapper{
+    background-color: rgba(244,239,239,255);
+}
