+++ src/components/order/PreviousOrders.css
@@ -37,6 +37,7 @@
 aside {
     margin: 20px;
     padding: 20px;
-    box-shadow: inset 0 0 0 2px rgba(0, 67, 50, 255);
+    box-shadow: inset 0 0 0 2px rgba(215,215,215,255);
     border-radius: 10px;
+    background-color: white;
 }
