+++ src/pages/common/404Page.css
@@ -1,0 +1,3 @@
+.page-not-found-container{
+    min-height: 44vh;
+}
