+++ src/pages/home/Home.css
@@ -27,7 +27,7 @@
     display: flex;
     align-items: center;
     justify-content: center;
-    color: rgba(192, 161, 89, 255);
+    color: white;
     font-size: 15px;
     border-radius: 8px;
     text-decoration: none;
