+++ src/pages/order/AddOrder.css
@@ -2,3 +2,4 @@
     display: flex;
     justify-content: space-between;
 }
+
