+++ src/pages/order/Orders.css
@@ -1,7 +1,3 @@
-.order-table-container {
-  margin: 20px;
-}
-
 .order-table {
   width: 100%;
   border-collapse: collapse;
@@ -9,9 +5,8 @@
 }
 
 .order-header {
-  background-color: #f2f2f2;
-  background-color: rgba(0, 67, 50, 255);
-  color: rgba(192, 161, 89, 255);
+  background-color:  rgba(245,245,245,255);
+  color: rgba(54,54,54,255);
   border: 1px solid #ddd;
   padding: 8px;
   text-align: left;
@@ -31,4 +26,8 @@
   border: none;
   border-radius: 4px;
   cursor: pointer;
+}
+
+.orders-heading{
+  color: rgba(54,54,54,255);
 }
